Split DataSourceStatusRegistryService (#11310)
* Remove duplicated methods

* Remove duplicated methods

* Rename RegistryService to PersistService

* Add subscriber package

* Split DataSourceStatusRegistryService
