+++ shardingsphere-governance/shardingsphere-governance-core/src/main/java/org/apache/shardingsphere/governance/core/registry/RegistryCenter.java
@@ -22,17 +22,18 @@ import org.apache.shardingsphere.governance.core.GovernanceInstance;
 import org.apache.shardingsphere.governance.core.lock.service.LockRegistryService;
 import org.apache.shardingsphere.governance.core.registry.cache.subscriber.ScalingRegistrySubscriber;
 import org.apache.shardingsphere.governance.core.registry.config.service.impl.DataSourcePersistService;
-import org.apache.shardingsphere.governance.core.registry.config.service.impl.DataSourceRegistrySubscriber;
 import org.apache.shardingsphere.governance.core.registry.config.service.impl.GlobalRulePersistService;
-import org.apache.shardingsphere.governance.core.registry.config.service.impl.GlobalRuleRegistrySubscriber;
 import org.apache.shardingsphere.governance.core.registry.config.service.impl.PropertiesPersistService;
 import org.apache.shardingsphere.governance.core.registry.config.service.impl.SchemaRulePersistService;
-import org.apache.shardingsphere.governance.core.registry.config.service.impl.SchemaRuleRegistrySubscriber;
+import org.apache.shardingsphere.governance.core.registry.config.subscriber.DataSourceRegistrySubscriber;
+import org.apache.shardingsphere.governance.core.registry.config.subscriber.GlobalRuleRegistrySubscriber;
+import org.apache.shardingsphere.governance.core.registry.config.subscriber.SchemaRuleRegistrySubscriber;
 import org.apache.shardingsphere.governance.core.registry.metadata.service.SchemaRegistryService;
 import org.apache.shardingsphere.governance.core.registry.process.subscriber.ProcessRegistrySubscriber;
 import org.apache.shardingsphere.governance.core.registry.state.node.StatesNode;
 import org.apache.shardingsphere.governance.core.registry.state.service.DataSourceStatusRegistryService;
 import org.apache.shardingsphere.governance.core.registry.state.service.UserStatusRegistryService;
+import org.apache.shardingsphere.governance.core.registry.state.subscriber.DataSourceStatusRegistrySubscriber;
 import org.apache.shardingsphere.governance.repository.spi.RegistryCenterRepository;
 import org.apache.shardingsphere.infra.config.RuleConfiguration;
 import org.apache.shardingsphere.infra.config.datasource.DataSourceConfiguration;
@@ -83,13 +84,14 @@ public final class RegistryCenter {
         schemaService = new SchemaRegistryService(repository);
         dataSourceStatusService = new DataSourceStatusRegistryService(repository);
         lockService = new LockRegistryService(repository);
-        createSubscriber(repository);
+        createSubscribers(repository);
     }
     
-    private void createSubscriber(final RegistryCenterRepository repository) {
+    private void createSubscribers(final RegistryCenterRepository repository) {
         new DataSourceRegistrySubscriber(dataSourceService);
         new GlobalRuleRegistrySubscriber(globalRuleService, new UserStatusRegistryService(repository));
         new SchemaRuleRegistrySubscriber(schemaRuleService);
+        new DataSourceStatusRegistrySubscriber(repository);
         new ScalingRegistrySubscriber(repository, schemaRuleService);
         new ProcessRegistrySubscriber(repository);
         ShardingSphereEventBus.getInstance().register(this);
