+++ shardingsphere-governance/shardingsphere-governance-core/src/main/java/org/apache/shardingsphere/governance/core/registry/state/service/DataSourceStatusRegistryService.java
@@ -18,13 +18,10 @@
 package org.apache.shardingsphere.governance.core.registry.state.service;
 
 import com.google.common.base.Strings;
-import com.google.common.eventbus.Subscribe;
+import lombok.RequiredArgsConstructor;
 import org.apache.shardingsphere.governance.core.registry.state.ResourceState;
 import org.apache.shardingsphere.governance.core.registry.state.node.StatesNode;
 import org.apache.shardingsphere.governance.repository.spi.RegistryCenterRepository;
-import org.apache.shardingsphere.infra.eventbus.ShardingSphereEventBus;
-import org.apache.shardingsphere.infra.rule.event.impl.DataSourceDisabledEvent;
-import org.apache.shardingsphere.infra.rule.event.impl.PrimaryDataSourceEvent;
 
 import java.util.Collection;
 import java.util.stream.Collectors;
@@ -32,14 +29,10 @@ import java.util.stream.Collectors;
 /**
  * Data source status registry service.
  */
+@RequiredArgsConstructor
 public final class DataSourceStatusRegistryService {
     
     private final RegistryCenterRepository repository;
-    
-    public DataSourceStatusRegistryService(final RegistryCenterRepository repository) {
-        this.repository = repository;
-        ShardingSphereEventBus.getInstance().register(this);
-    }
     
     /**
      * Load disabled data source names.
@@ -58,26 +51,5 @@ public final class DataSourceStatusRegistryService {
     
     private String getDataSourceNodeData(final String schemaName, final String dataSourceName) {
         return repository.get(StatesNode.getDataSourcePath(schemaName, dataSourceName));
-    }
-    
-    /**
-     * Update data source disabled state.
-     *
-     * @param event data source disabled event
-     */
-    @Subscribe
-    public void update(final DataSourceDisabledEvent event) {
-        String value = event.isDisabled() ? ResourceState.DISABLED.toString() : "";
-        repository.persist(StatesNode.getDataSourcePath(event.getSchemaName(), event.getDataSourceName()), value);
-    }
-    
-    /**
-     * Update primary data source state.
-     *
-     * @param event primary data source event
-     */
-    @Subscribe
-    public void update(final PrimaryDataSourceEvent event) {
-        repository.persist(StatesNode.getPrimaryDataSourcePath(event.getSchemaName(), event.getGroupName()), event.getDataSourceName());
     }
 }
